+++ server/src/validators/createUser.validator.ts
@@ -1,0 +1,9 @@
+import { checkSchema } from "express-validator";
+
+const userSchema = {
+  username: {},
+};
+
+const validateCreateUser = checkSchema(userSchema, ["body"]);
+
+export default validateCreateUser;