+++ server/src/validators/params.validator.ts
@@ -1,0 +1,32 @@
+import asyncHandler from "express-async-handler";
+import { checkSchema, validationResult, Schema } from "express-validator";
+import { getUser } from "../services/user";
+
+const isIdValid = async (id: string) => {
+  const userExists = !!(await getUser(id));
+  if (!userExists)
+    return Promise.reject(`User with the id, ${id}, does not exist`);
+};
+
+const userSchema = {
+  id: {
+    trim: true,
+    custom: {
+      options: isIdValid,
+    },
+    escape: true,
+  },
+};
+
+const validateParams = (schema: Schema) => {
+  return asyncHandler(async (req, res, next) => {
+    await checkSchema(schema, ["params"]).run(req);
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) next({ status: 404, message: "Resource not found" });
+
+    next();
+  });
+};
+
+export { validateParams as default, userSchema };