+++ server/src/validators/validators.ts
@@ -1,0 +1,3 @@
+import validateCreateUser from "./createUser.validator";
+
+export { validateCreateUser };